requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
pytz==2024.1
python-telegram-bot==13.15
requests==2.31.0
python-dotenv>=0.21.0   # чтобы Flask и бот могли читать .env
numpy==1.24.3
Flask-Limiter==2.8.0
Flask-Mail==0.9.1
